:truck: :recycle: DB로직 DB폴더 내 DAO폴더내로 옮김
