+++ app/src/main/java/woowacourse/omok/model/omokGame/Board.kt
@@ -5,7 +5,7 @@ import woowacourse.omok.model.board.CoordsNumber
 import woowacourse.omok.model.board.Position
 import woowacourse.omok.model.board.Stone
 
-class Board(val gameBoard: Array<Array<Stone>> = Array(BOARD_SIZE) { Array(BOARD_SIZE) { Stone.EMPTY } }) {
+class Board(var gameBoard: Array<Array<Stone>> = Array(BOARD_SIZE) { Array(BOARD_SIZE) { Stone.EMPTY } }) {
     private var renjuGameRule: GameRuleAdapter = GameRuleAdapter()
 
     init {
